+++ solarnet/datum/src/main/java/net/solarnetwork/central/datum/biz/dao/DaoDatumMaintenanceBiz.java
@@ -1,21 +1,21 @@
 /* ==================================================================
  * DaoDatumMaintenanceBiz.java - 10/04/2019 9:03:02 am
- * 
+ *
  * Copyright 2019 SolarNetwork.net Dev Team
- * 
- * This program is free software; you can redistribute it and/or 
- * modify it under the terms of the GNU General Public License as 
- * published by the Free Software Foundation; either version 2 of 
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License as
+ * published by the Free Software Foundation; either version 2 of
  * the License, or (at your option) any later version.
- * 
- * This program is distributed in the hope that it will be useful, 
- * but WITHOUT ANY WARRANTY; without even the implied warranty of 
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU 
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
  * General Public License for more details.
- * 
- * You should have received a copy of the GNU General Public License 
- * along with this program; if not, write to the Free Software 
- * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA
  * 02111-1307 USA
  * ==================================================================
  */
@@ -40,17 +40,17 @@ import net.solarnetwork.central.datum.domain.StaleAggregateDatum;
 import net.solarnetwork.central.datum.v2.dao.BasicDatumCriteria;
 import net.solarnetwork.central.datum.v2.dao.DatumMaintenanceDao;
 import net.solarnetwork.central.datum.v2.dao.DatumStreamMetadataDao;
+import net.solarnetwork.central.datum.v2.support.DatumUtils;
+import net.solarnetwork.central.domain.FilterResults;
+import net.solarnetwork.central.support.BasicFilterResults;
+import net.solarnetwork.domain.SortDescriptor;
 import net.solarnetwork.domain.datum.DatumStreamMetadata;
 import net.solarnetwork.domain.datum.ObjectDatumKind;
 import net.solarnetwork.domain.datum.ObjectDatumStreamMetadata;
-import net.solarnetwork.central.datum.v2.support.DatumUtils;
-import net.solarnetwork.central.domain.FilterResults;
-import net.solarnetwork.domain.SortDescriptor;
-import net.solarnetwork.central.support.BasicFilterResults;
 
 /**
  * DAO based implementation of {@link DatumMaintenanceBiz}.
- * 
+ *
  * @author matt
  * @version 2.0
  * @since 1.6
@@ -64,7 +64,7 @@ public class DaoDatumMaintenanceBiz implements DatumMaintenanceBiz {
 
 	/**
 	 * Constructor.
-	 * 
+	 *
 	 * @param datumDao
 	 *        the datum DAO to use
 	 * @param metaDao
@@ -90,7 +90,7 @@ public class DaoDatumMaintenanceBiz implements DatumMaintenanceBiz {
 	@Transactional(readOnly = true, propagation = Propagation.SUPPORTS)
 	@Override
 	public FilterResults<StaleAggregateDatum> findStaleAggregateDatum(GeneralNodeDatumFilter criteria,
-			List<SortDescriptor> sortDescriptors, Integer offset, Integer max) {
+			List<SortDescriptor> sortDescriptors, Long offset, Integer max) {
 		BasicDatumCriteria c = DatumUtils.criteriaFromFilter(criteria, sortDescriptors, offset, max);
 		c.setObjectKind(ObjectDatumKind.Node);
 		DatumUtils.populateAggregationType(criteria, c);
